requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
requests==2.26.0
python-dotenv==0.19.0
pymysql==1.0.2
cryptography==35.0.0
pydantic==1.10.7
jieba==0.42.1
pycryptodomex==3.17.0 